Renaming files and small changes as per PR comments
