+++ src/parameters/all_parameters.cpp
@@ -117,7 +117,7 @@ void AllParameters::declare_parameters (dealii::ParameterHandler &prm)
                       " dual_weighted_residual_mesh_adaptation | "
                       " taylor_green_vortex_energy_check | "
                       " taylor_green_vortex_restart_check | "
-                      " time_refinement_study_advection"),
+                      " time_refinement_study"),
                       "The type of test we want to solve. "
                       "Choices are (only run control has been coded up for now)" 
                       " <run_control | " 
@@ -145,7 +145,7 @@ void AllParameters::declare_parameters (dealii::ParameterHandler &prm)
                       "  dual_weighted_residual_mesh_adaptation | "
                       "  taylor_green_vortex_energy_check | "
                       "  taylor_green_vortex_restart_check | "
-                      "  time_refinement_study_advection>.");
+                      "  time_refinement_study>.");
 
     prm.declare_entry("pde_type", "advection",
                       dealii::Patterns::Selection(
@@ -241,7 +241,7 @@ void AllParameters::parse_parameters (dealii::ParameterHandler &prm)
     else if (test_string == "dual_weighted_residual_mesh_adaptation")   { test_type = dual_weighted_residual_mesh_adaptation; }
     else if (test_string == "taylor_green_vortex_energy_check")         { test_type = taylor_green_vortex_energy_check; }
     else if (test_string == "taylor_green_vortex_restart_check")        { test_type = taylor_green_vortex_restart_check; }
-    else if (test_string == "time_refinement_study_advection")          { test_type = time_refinement_study_advection; }
+    else if (test_string == "time_refinement_study")                    { test_type = time_refinement_study; }
 
     const std::string pde_string = prm.get("pde_type");
     if (pde_string == "advection") {
