+++ src/parameters/all_parameters.h
@@ -131,7 +131,7 @@ public:
         dual_weighted_residual_mesh_adaptation,
         taylor_green_vortex_energy_check,
         taylor_green_vortex_restart_check,
-        time_refinement_study_advection,
+        time_refinement_study,
     };
     TestType test_type; ///< Selected TestType from the input file.
