+++ src/testing/CMakeLists.txt
@@ -31,7 +31,7 @@ set(TESTS_SOURCE
     taylor_green_vortex_energy_check.cpp
     taylor_green_vortex_restart_check.cpp
     flow_solver_cases/naca0012.cpp
-    time_refinement_study_advection.cpp)
+    time_refinement_study.cpp)
 
 foreach(dim RANGE 1 3)
     # Output library
