+++ tests/integration_tests_control_files/time_refinement_study/time_refinement_study_advection_explicit.prm
@@ -3,7 +3,7 @@
 # Number of dimensions
 
 set dimension = 1 
-set test_type = time_refinement_study_advection
+set test_type = time_refinement_study
 set pde_type = advection
 
 # Note: this was added to turn off check_same_coords() -- has no other function when dim!=1
